lint

+++ test/concrete/RainterpreterExpressionDeployer.touchDeployer.t.sol
@@ -12,7 +12,7 @@ contract RainterpreterExpressionDeployerTouchDeployer is RainterpreterExpression
         vm.expectCall(
             address(iDeployer),
             abi.encodeWithSelector(
-                IExpressionDeployerV1.deployExpression.selector, new bytes[](0), new uint256[](0), new uint8[](0)
+                IExpressionDeployerV1.deployExpression.selector, new bytes[](0), new uint256[](0), new uint256[](0)
             ),
             1
         );
